Outsourcing classes and pylint, mypy corrections.
